+++ zhmcclient/_certificates.py
@@ -265,7 +265,7 @@ class Certificate(BaseResource):
           :exc:`~zhmcclient.ConnectionError`
         """
         # pylint: disable=protected-access
-        self.manager.session.delete(self.uri)
+        self.manager.session.delete(self.uri, resource=self)
         self.manager._name_uri_cache.delete(
             self.get_properties_local(self.manager._name_prop, None))
 
@@ -298,7 +298,7 @@ class Certificate(BaseResource):
           :exc:`~zhmcclient.ConnectionError`
         """
         # pylint: disable=protected-access
-        self.manager.session.post(self.uri, body=properties)
+        self.manager.session.post(self.uri, resource=self, body=properties)
         is_rename = self.manager._name_prop in properties
         if is_rename:
             # Delete the old name from the cache
@@ -333,7 +333,7 @@ class Certificate(BaseResource):
         """
         # pylint: disable=protected-access
         return self.manager.session.get(
-            '{}/operations/get-encoded'.format(self.uri))
+            '{}/operations/get-encoded'.format(self.uri), resource=self)
 
     def dump(self):
         """