+++ zhmcclient/_user.py
@@ -264,7 +264,7 @@ class User(BaseResource):
           :exc:`~zhmcclient.ConnectionError`
         """
         # pylint: disable=protected-access
-        self.manager.session.delete(self.uri)
+        self.manager.session.delete(self.uri, resource=self)
         self.manager._name_uri_cache.delete(
             self.get_properties_local(self.manager._name_prop, None))
 
@@ -301,7 +301,7 @@ class User(BaseResource):
           :exc:`~zhmcclient.ConnectionError`
         """
         # pylint: disable=protected-access
-        self.manager.session.post(self.uri, body=properties)
+        self.manager.session.post(self.uri, resource=self, body=properties)
 
         # The name of Users cannot be updated. An attempt to do so should cause
         # HTTPError to be raised in the POST above, so we assert that here,
@@ -337,7 +337,7 @@ class User(BaseResource):
             'user-role-uri': user_role.uri
         }
         self.manager.session.post(
-            self.uri + '/operations/add-user-role',
+            self.uri + '/operations/add-user-role', resource=self,
             body=body)
 
     @logged_api_call
@@ -368,5 +368,5 @@ class User(BaseResource):
             'user-role-uri': user_role.uri
         }
         self.manager.session.post(
-            self.uri + '/operations/remove-user-role',
+            self.uri + '/operations/remove-user-role', resource=self,
             body=body)